Move InfoStream and OutputManager to Platforms/Host
