+++ src/Platforms/Host/InfoStream.h
