+++ src/QMCDrivers/tests/test_WalkerControlMPI.cpp
@@ -21,7 +21,7 @@
 #include "QMCDrivers/QMCDriverInput.h"
 #include "QMCDrivers/tests/test_WalkerControlMPI.h"
 #include "Utilities/MPIExceptionWrapper.hpp"
-#include "Utilities/OutputManager.h"
+#include "Platforms/Host/OutputManager.h"
 
 
 //#include "Concurrency/Info.hpp"
